-see milestone "Release v0.5.2 Graph State Re-Work" for specific issues. https://github.com/terrameta/cloudgraph/issues?milestone=3

+++ cloudgraph-common/src/main/java/org/cloudgraph/common/key/GraphStatefullColumnKeyFactory.java
@@ -51,5 +51,5 @@ public interface GraphStatefullColumnKeyFactory {
 	 * @return the column key bytes
 	 */
 	public byte[] createColumnKey(PlasmaType type, 
-			Long dataObjectSeqNum, PlasmaProperty property);
+			Integer dataObjectSeqNum, PlasmaProperty property);
 }
